Add lib.rs to be able to use it in codebase
